[Updated] Publish 1 patch 1 Clean Base

+++ MMOCoreORB/src/templates/tangible/SharedWeaponObjectTemplate.cpp
@@ -216,7 +216,7 @@ void SharedWeaponObjectTemplate::readObject(LuaObject* templateData) {
 	healthAttackCost = templateData->getIntField("healthAttackCost");
 	actionAttackCost = templateData->getIntField("actionAttackCost");
 	mindAttackCost = templateData->getIntField("mindAttackCost");
-	forceCost = templateData->getFloatField("forceCost");
+	forceCost = templateData->getIntField("forceCost");
 
 	pointBlankAccuracy = templateData->getIntField("pointBlankAccuracy");
 	pointBlankRange = templateData->getIntField("pointBlankRange");
